Implement Fading for Audio Cue Backend
This allows audio cues to be faded to any value using a curve function.
The way this is currently implemented changes the system volume for some
reason. DON'T RUN CUES ABOVE 1.0 TARGET VOLUME. According to the documentation,
the playbin can have a volume of up to 10, which is 1000% (yes, 1000%).
I need to research how to adjust the volume relative to the source still.

Also, fading is only used for the panic button (Default: 500ms) but should
be easy to implement elsewhere.