Fix GUI updating and add Object Cache
GUI Updates (progress bars) used to be updated with a clock spawned
by gstreamer. This had the odd tendency of crashing the python
process if the update function was called too frequently. Switching
to GLib.timeout_add allows the function to be called more often and
looks better.

Unfortunately, this seems to have broken fading for some reason...

Also, work has started on the object cache, which will allow cues
to reference each other when writing to disk (autofollow targets,
automation cues, etc.)