Cache audio file duration and load asyncronously
This should significantly speed up loading times. It seems however
that `GstPbutils.Discoverer` doesn't return reliable tiems for
certain file types (It tends to be off by a few parts of a second),
which can lead to a large number of objects being stored in the
repository.

Initial work on #13