Refocus cue list after clicking on buttons
Previously, GTK would grab the keyboard input focus when a user
clicked a button and direct it towards that button, forcing the user
to have to click the cue list becore they could trigger a cue with
the keyboard. This change grabs the focus of the cue list widget
after a button is clicked.

Fixes #14