Limit duration difference warning within a small value epsilon
When the duration of an audio file is discovered, it is compared to
the cached value. For some file formats, GTK seems to get the duration
within +/- 5ms. Don't warn or update the cached duration unless it is
more than some small value epsilon. This is configurable in the project
properties window.

Fixes #13